added Mappers for some classes
