+++ src/main/java/com/example/springsecurityfivecass/service/TeacherService.java
@@ -1,7 +1,7 @@
 package com.example.springsecurityfivecass.service;
 
 import com.example.springsecurityfivecass.entity.Teacher;
-import com.example.springsecurityfivecass.repository.CourseRepository;
+import com.example.springsecurityfivecass.entity.mappers.TeacherMappers;
 import com.example.springsecurityfivecass.repository.TeacherRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,14 +13,14 @@ import java.util.List;
 public class TeacherService {
 
     private final TeacherRepository teacherRepository;
-    private final CourseRepository courseRepository;
+    private final TeacherMappers teacherMappers;
 
     public List<Teacher> getAllTeacher(){
         return teacherRepository.findAll();
     }
 
     public Teacher save(Teacher teacher){
-        Teacher teacher1 = mapToEntity(teacher);
+        Teacher teacher1 = teacherMappers.mapToEntity(teacher);
         teacherRepository.save(teacher1);
         return teacher1;
     }
@@ -38,13 +38,4 @@ public class TeacherService {
         teacherRepository.deleteById(id);
     }
 
-    public Teacher mapToEntity(Teacher teacher){
-        Teacher teacher1 = new Teacher();
-        teacher1.setFirstName(teacher.getFirstName());
-        teacher1.setLastName(teacher.getLastName());
-        teacher1.setEmail(teacher.getEmail());
-        teacher1.setCourse(courseRepository.getById(teacher.getCourse().getId()));
-        return teacher1;
-
-    }
 }
